+++ arenatacticsserver/scenes/UI/characters/character_select_screen/3d_scene/character_select_3d.gd
@@ -7,18 +7,18 @@ class_name CharacterSelect3D
 func _ready():
 	pass
 
-func build(character: CharacterInstance):
+func build(class_def: ClassDefinition):
 	reset_diorama()
-	spawn_diorama(character)
+	spawn_diorama(class_def)
 
 func reset_diorama():
 	# fade anim here
 	# camera movement/blur ?
 	for child in chara_spawn.get_children():
-		child.free()
+		child.queue_free()
 
-func spawn_diorama(character: CharacterInstance):
+func spawn_diorama(class_def: ClassDefinition):
 	# fade in here/play fade out in reverse
-	var new_instance = ClassLoader.get_class_instance(character.character_class)
+	var new_instance = class_def.instantiate()
 	print_debug("Spawing %s on Diorama" % new_instance.character_class)
 	chara_spawn.add_child(new_instance)