+++ design_docs/history.md
@@ -27,4 +27,4 @@
 
 ## 05/02/2025
 
-- Class Character Instanciation 9-10
+- Class Character Instanciation 9-10 10-11